requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
chromadb
google-generativeai
pdfminer.six
jinja2
# NLP & Machine Learning
torch
transformers
sentence-transformers
numpy
pandas
